+++ src/native.h
@@ -5,6 +5,14 @@
 #include "object.h"
 #include "vm.h"
 
+typedef struct {
+  const char *name;
+  NativeFn function;
+} NativeDefinition;
+
+extern const NativeDefinition nativeDefinitions[];
+extern const int nativeDefinitionCount;
+
 void defineNative(VM *vm, const char *name, NativeFn function);
 void defineAllNatives(VM *vm);